// ai-chat-frontend/ai-chat-frontend/src/app/shared/models/websocket/websocket-action.ts
export enum WebsocketAction {
  GET_CONVERSATIONS = 'GET_CONVERSATIONS',
  CREATE_AI_MESSAGE = 'CREATE_AI_MESSAGE',
  CREATE_AI_CHAT_WITH_INIT_MESSAGE = 'CREATE_AI_CHAT_WITH_INIT_MESSAGE',
  ADD_USER_MESSAGE = 'ADD_USER_MESSAGE',
  DELETE_CONVERSATION = 'DELETE_CONVERSATION',
  GET_CONVERSATION_BY_ID = 'GET_CONVERSATION_BY_ID',
  ADD_USER_TO_MESSAGE = 'ADD_USER_TO_MESSAGE',
  REMOVE_USER_FROM_MESSAGE = 'REMOVE_USER_FROM_MESSAGE',
  SEARCH_USER = 'SEARCH_USER'

}